refactor: Move some rental detail components to common package and rename components (#64)
