+++ app/src/main/java/com/example/rentit/common/component/RentalSummary.kt
@@ -1,4 +1,4 @@
-package com.example.rentit.presentation.rentaldetail.common.components
+package com.example.rentit.common.component
 
 import android.os.Build
 import androidx.annotation.RequiresApi
