+++ app/src/main/java/com/example/rentit/presentation/rentaldetail/renter/components/PaidContent.kt
@@ -11,9 +11,9 @@ import com.example.rentit.common.enums.RentalStatus
 import com.example.rentit.common.theme.Gray400
 import com.example.rentit.common.theme.RentItTheme
 import com.example.rentit.presentation.rentaldetail.common.model.PriceItemUiModel
-import com.example.rentit.presentation.rentaldetail.common.components.section.PaymentInfoSection
+import com.example.rentit.presentation.rentaldetail.common.components.section.RentalPaymentSection
 import com.example.rentit.presentation.rentaldetail.common.components.section.RentalInfoSection
-import com.example.rentit.presentation.rentaldetail.common.components.section.TrackingInfoSection
+import com.example.rentit.presentation.rentaldetail.common.components.section.RentalTrackingSection
 import com.example.rentit.presentation.rentaldetail.common.model.RentalSummaryUiModel
 import com.example.rentit.presentation.rentaldetail.renter.model.RentalStatusRenterUiModel
 
@@ -49,13 +49,13 @@ fun PaidContent(
         rentalInfo = paidData.rentalSummary,
     )
 
-    PaymentInfoSection(
+    RentalPaymentSection(
         title = stringResource(R.string.screen_rental_detail_renter_paid_price_title),
         priceItems = priceItems,
         totalLabel = stringResource(R.string.screen_rental_detail_renter_paid_price_label_total)
     )
 
-    TrackingInfoSection(
+    RentalTrackingSection(
         rentalTrackingNumber = paidData.rentalTrackingNumber,
     )
 }
