+++ app/src/main/java/com/example/rentit/presentation/rentaldetail/renter/components/RentingContent.kt
@@ -29,10 +29,10 @@ import com.example.rentit.common.theme.Gray100
 import com.example.rentit.common.theme.RentItTheme
 import com.example.rentit.presentation.rentaldetail.common.model.PriceItemUiModel
 import com.example.rentit.presentation.rentaldetail.common.components.NoticeBanner
-import com.example.rentit.presentation.rentaldetail.common.components.section.PaymentInfoSection
+import com.example.rentit.presentation.rentaldetail.common.components.section.RentalPaymentSection
 import com.example.rentit.presentation.rentaldetail.common.components.section.RentalInfoSection
-import com.example.rentit.presentation.rentaldetail.common.components.section.TaskSection
-import com.example.rentit.presentation.rentaldetail.common.components.section.TrackingInfoSection
+import com.example.rentit.presentation.rentaldetail.common.components.section.RentalTaskSection
+import com.example.rentit.presentation.rentaldetail.common.components.section.RentalTrackingSection
 import com.example.rentit.presentation.rentaldetail.common.model.RentalSummaryUiModel
 import com.example.rentit.presentation.rentaldetail.renter.model.RentalStatusRenterUiModel
 import com.example.rentit.presentation.rentaldetail.renter.model.RentingStatus
@@ -72,7 +72,7 @@ fun RentingContent(
         rentalInfo = rentingData.rentalSummary,
     )
 
-    TaskSection(
+    RentalTaskSection(
         title = stringResource(R.string.screen_rental_detail_renter_return_task_title),
         guideText = stringResource(R.string.screen_rental_detail_renter_return_task_info),
         policyText = stringResource(R.string.screen_rental_detail_renter_return_task_policy),
@@ -89,13 +89,13 @@ fun RentingContent(
         }
     }
 
-    PaymentInfoSection(
+    RentalPaymentSection(
         title = stringResource(R.string.screen_rental_detail_renter_paid_price_title),
         priceItems = priceItems,
         totalLabel = stringResource(R.string.screen_rental_detail_renter_paid_price_label_total)
     )
 
-    TrackingInfoSection(
+    RentalTrackingSection(
         rentalTrackingNumber = rentingData.rentalTrackingNumber,
     )
 }
